separated BulkClone and BulkDelete into their own operation traits
